Refactored folder structure, no more networkView
