+++ src/components/sidebar/linkView/LinkView.tsx
@@ -8,6 +8,7 @@ import { FiChevronRight, FiChevronLeft } from 'react-icons/fi';
 import ButtonType from '~/enums/buttonType';
 import { IValidationResult } from '~/validation/FormValidator';
 import ViewFormBase from '~/components/shared/inheritedComponents/ViewFormBase';
+import Loader from '~/components/shared/loader/Loader';
 import LinkService from '~/services/linkService';
 import nodeTypeCodeList from '~/codeLists/nodeTypeCodeList';
 import linkValidationModel from '~/validation/models/linkValidationModel';
@@ -20,11 +21,10 @@ import navigator from '~/routing/navigator';
 import { LinkStore } from '~/stores/linkStore';
 import { MapStore } from '~/stores/mapStore';
 import { ErrorStore } from '~/stores/errorStore';
-import { Dropdown, Button, TransitToggleButtonBar } from '../../../controls';
-import InputContainer from '../../InputContainer';
-import TextContainer from '../../TextContainer';
-import Loader from '../../../shared/loader/Loader';
-import SidebarHeader from '../../SidebarHeader';
+import { Dropdown, Button, TransitToggleButtonBar } from '../../controls';
+import InputContainer from '../InputContainer';
+import TextContainer from '../TextContainer';
+import SidebarHeader from '../SidebarHeader';
 import * as s from './linkView.scss';
 
 interface ILinkViewState {
