+++ src/components/sidebar/linkView/multiTabTextarea.scss
@@ -1,5 +1,5 @@
-@import '../../../shared/styles/common';
-@import '../../../shared/styles/flexbox';
+@import '../../shared/styles/common';
+@import '../../shared/styles/flexbox';
 
 .tabsInputContainer {
     padding-top: 10px;
