--- src/components/sidebar/networkView/NetworkDefault.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import * as s from './networkView.scss';
-
-class NetworkDefault extends React.Component {
-
-    public render() {
-        return (
-            <div className={s.networkView}>
-                <h2>Verkon muokkaus</h2>
-                <p>TODO: Search functionality</p>
-            </div>
-        );
-    }
-}
-
-export default NetworkDefault;
