--- src/components/sidebar/networkView/networkView.scss
@@ -1,4 +1,0 @@
-.networkView {
-    width: 500px;
-    overflow-y: auto;
-}
