--- src/components/sidebar/networkView/nodeView/stopForm.scss
@@ -1,11 +1,0 @@
-@import '../../../shared/styles/common.scss';
-@import '../../../shared/styles/flexbox';
-@import '../../../shared/styles/form';
-
-.stopView {
-  margin-top: 15px;
-  
-  .sectionHeader {
-    margin-top: 15px;
-  }  
-}
