+++ src/components/sidebar/nodeView/NodeCoordinatesListView.tsx
