+++ src/components/sidebar/nodeView/nodeCoordinatesListView.scss
@@ -1,4 +1,4 @@
-@import '../../../shared/styles/common';
+@import '../../shared/styles/common';
 
 .nodeCoordinatesListView {
   .measured {
