+++ src/components/sidebar/nodeView/nodeCoordinatesView.scss
@@ -1,4 +1,4 @@
-@import '../../../shared/styles/common';
+@import '../../shared/styles/common';
 
 .nodeCoordinatesView {
   padding-bottom: 10px;
