+++ src/components/sidebar/nodeView/nodeView.scss
@@ -1,6 +1,6 @@
-@import '../../../shared/styles/common.scss';
-@import '../../../shared/styles/flexbox';
-@import '../../../shared/styles/form';
+@import '../../shared/styles/common.scss';
+@import '../../shared/styles/flexbox';
+@import '../../shared/styles/form';
 
 .nodeView {
     display: flex;
