+++ src/components/sidebar/nodeView/stopForm.scss
@@ -1,0 +1,11 @@
+@import '../../shared/styles/common.scss';
+@import '../../shared/styles/flexbox';
+@import '../../shared/styles/form';
+
+.stopView {
+  margin-top: 15px;
+
+  .sectionHeader {
+    margin-top: 15px;
+  }
+}
